+++ src/mailman/docs/NEWS.rst
@@ -43,6 +43,8 @@ Bugs
  * Collapse multiple ``Re:`` in Subject headers.  Given by Mark Sapiro.
    (Closes: #147)
  * Added Trove classifiers to setup.py.  (Closes: #152)
+ * Fix the processing of subscription confirmation messages when the mailing
+   list is set to confirm-then-moderate.  (Closes #114)
 
 Configuration
 -------------